@@ -39,8 +39,8 @@
   "devDependencies": {
     "cross-env": "^5.1.3",
     "css-loader": "^0.28.9",
-    "electron": "1.7.12",
-    "electron-builder": "^19.55.3",
+    "electron": "1.8.2",
+    "electron-builder": "^19.56.0",
     "electron-webpack": "1.12.1",
     "electron-webpack-eslint": "^1.3.0",
     "electron-webpack-vue": "^1.3.0",
